MapsActivity updated to include building highlights
