--- app/src/main/java/com/conupods/GeoJsonActivity.java
@@ -1,114 +1,0 @@
-package com.conupods;
-
-import androidx.fragment.app.FragmentActivity;
-
-import android.os.AsyncTask;
-import android.os.Bundle;
-import android.util.Log;
-import android.widget.Toast;
-
-import com.conupods.R;
-import com.google.android.gms.maps.CameraUpdateFactory;
-import com.google.android.gms.maps.GoogleMap;
-import com.google.android.gms.maps.OnMapReadyCallback;
-import com.google.android.gms.maps.SupportMapFragment;
-import com.google.android.gms.maps.model.LatLng;
-import com.google.android.gms.maps.model.MarkerOptions;
-import com.google.maps.android.data.Feature;
-import com.google.maps.android.data.geojson.GeoJsonFeature;
-import com.google.maps.android.data.geojson.GeoJsonLayer;
-import com.google.maps.android.data.geojson.GeoJsonPolygonStyle;
-
-import org.json.JSONException;
-import org.json.JSONObject;
-
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.net.URL;
-
-public class GeoJsonActivity extends FragmentActivity implements OnMapReadyCallback {
-
-    private GoogleMap mMap;
-    private LatLng sgw = new LatLng(45.496080, -73.577957);
-    private LatLng loy = new LatLng(45.458333, -73.640450);
-    private LatLng current = sgw;
-
-    private final static String mLogTag = "GeoJsonOverlay";
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_maps);
-        // Obtain the SupportMapFragment and get notified when the map is ready to be used.
-        SupportMapFragment mapFragment = (SupportMapFragment) getSupportFragmentManager().findFragmentById(R.id.map);
-        mapFragment.getMapAsync(this);
-
-    }
-
-    @Override
-    public void onMapReady(GoogleMap googleMap) {
-        mMap = googleMap;
-        LatLng sydney = new LatLng(-34, 151);
-        mMap.moveCamera(CameraUpdateFactory.newLatLng(current));
-        retrieveFileFromUrl();
-
-    }
-
-    private void addColorsToMarkers(GeoJsonLayer layer) {
-        // Iterate over all the features stored in the layer
-        for (GeoJsonFeature feature : layer.getFeatures()) {
-            // Check if the  property exists
-            GeoJsonPolygonStyle buildingStyle = layer.getDefaultPolygonStyle();
-                buildingStyle.setFillColor(0x80eac700);
-                buildingStyle.setStrokeColor(0x80555555);
-                buildingStyle.setStrokeWidth(5);
-        }
-    }
-
-
-    private void retrieveFileFromUrl() {
-        new DownloadGeoJsonFile().execute(getString(R.string.geojson_url));
-    }
-
-    private void addGeoJsonLayerToMap(GeoJsonLayer layer) {
-        addColorsToMarkers(layer);
-        layer.addLayerToMap();
-    }
-    private class DownloadGeoJsonFile extends AsyncTask<String, Void, GeoJsonLayer> {
-
-        @Override
-        protected GeoJsonLayer doInBackground(String... params) {
-            try {
-                // Open a stream from the URL
-                InputStream stream = new URL(params[0]).openStream();
-                String line;
-                StringBuilder result = new StringBuilder();
-                BufferedReader reader = new BufferedReader(new InputStreamReader(stream));
-                while ((line = reader.readLine()) != null) {
-                    // Read and save each line of the stream
-                    result.append(line);
-                }
-                // Close the stream
-                reader.close();
-                stream.close();
-                return new GeoJsonLayer(mMap, new JSONObject(result.toString()));
-            } catch (IOException e) {
-                Log.e(mLogTag, "GeoJSON file could not be read");
-            } catch (JSONException e) {
-                Log.e(mLogTag, "GeoJSON file could not be converted to a JSONObject");
-            }
-            return null;
-        }
-        @Override
-        protected void onPostExecute(GeoJsonLayer layer) {
-            if (layer != null) {
-                addGeoJsonLayerToMap(layer);
-            }
-        }
-    }
-
-
-
-}
